libkcapi: switch to Amazon Linux 2023 as upstream
This aligns libkcapi with the version specified in the FIPS 140-3
security policy for the Amazon Linux 2023 Kernel Cryptographic API.

Bump the epoch to ensure that the "older" 1.4.0 version is preferred
over the "newer" 1.5.0 version from past core kit releases.

Trim the set of installed files down to just the ones referenced by
the security policy: `sha512hmac` and `libkcapi.so.1.4.0`.

Signed-off-by: Ben Cressey <bcressey@amazon.com>