sundog: Parse settings generators as a table
Earlier we used to have setting generator as a string, but we have now
changed it to a struct containing the command, strength and
skip-if-populated.

Hence after requesting the settings generators from api, we need to
change these in to Setting generator struct instance. We will use default
strength strong and skip_if_populated true if the setting generator
is a string and use what has been provided in the response otherwise.

Then after running the setting generator, we will send the weak and
strong setting settings separately to process them using api.