migrator: remove all the weak setting and settings-generator
We will remove the weak settings and all the metadata using the
migrator of the destination migrator.

The storewolf do not repopulate any metadata or setting, if it is
already present. As migrator runs before storewolf, if we will delete the
weak settings and all metadata in migrator, storewolf can populate
the setting-generator from defaults and sundog will populate the new
source using the new setting generator from default.

We are deleting all the metadata as we populate that using defaults only
and also the same is not possible to set the using API.