migrator: remove all the weak setting and settings-generator
We will remove the weak settings and settings genrators using the
migrator of the destination migrator.
The storewolf do not repopulate any metadata or setting, if it is
already presnt. As migrator runs before storewolf, if we will delete the
weak settings and setting-generators in migrator, storewolf can populate
the setting-generator from defaults and ssundog will populate the new
source using the new setting generator from default.